requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
redis>=5.0.0,<6.0.0
aiohttp>=3.9.0,<4.0.0
pillow>=10.0.0,<11.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
alembic>=1.13.0,<2.0.0
structlog>=24.1.0,<25.0.0
pytest>=8.0.0,<9.0.0
httpx>=0.27.0,<0.28.0
